deleteComment 200; 400; 404
