+++ controllers/reviews.controller.js
@@ -1,8 +1,6 @@
 const {
   selectAllReviews,
   selectReviewByID,
-  selectCommentsByReviewID,
-  insertComment,
   updateReviewVotes,
 } = require("../models/reviews.model");
 
@@ -20,26 +18,6 @@ exports.getReviewByID = (req, res, next) => {
   selectReviewByID(review_id)
     .then((review) => {
       res.status(200).send({ review });
-    })
-    .catch(next);
-};
-
-exports.getCommentsByReviewID = (req, res, next) => {
-  const review_id = req.params.review_id;
-  selectCommentsByReviewID(review_id)
-    .then((review) => {
-      res.status(200).send({ review });
-    })
-    .catch(next);
-};
-
-exports.postComment = (req, res, next) => {
-  const review_id = req.params.review_id;
-  const newComment = req.body;
-
-  insertComment(review_id, newComment)
-    .then((postedComment) => {
-      res.status(201).send({ postedComment });
     })
     .catch(next);
 };
