+++ models/reviews.model.js
@@ -70,45 +70,6 @@ exports.selectReviewByID = (review_id) => {
   }
 };
 
-exports.selectCommentsByReviewID = (review_id) => {
-  if (isNaN(review_id)) {
-    return Promise.reject({ status: 400, msg: "400: Bad Request" });
-  } else {
-    return checkExists("reviews", "review_id", review_id)
-      .then(() => {
-        return db.query(
-          `SELECT * FROM comments WHERE review_id = $1 ORDER BY created_at DESC;`,
-          [review_id]
-        );
-      })
-      .then((result) => {
-        return result.rows;
-      });
-  }
-};
-
-exports.insertComment = (review_id, newComment) => {
-  if (
-    !Object.keys(newComment).includes("username", "body") ||
-    !newComment.body ||
-    !newComment.username ||
-    isNaN(review_id)
-  ) {
-    return Promise.reject({ status: 400, msg: "400: Bad Request" });
-  } else {
-    return Promise.all([
-      checkExists("reviews", "review_id", review_id),
-      checkExists("users", "username", newComment.username),
-    ]).then(() => {
-      let queryStr = `INSERT INTO comments (review_id, author, body) VALUES ($1, $2, $3) RETURNING *;`;
-      const queryVals = [review_id, newComment.username, newComment.body];
-      return db.query(queryStr, queryVals).then((result) => {
-        return result.rows[0];
-      });
-    });
-  }
-};
-
 exports.updateReviewVotes = (review_id, patchContent) => {
   if (
     !patchContent ||
